(지희) join 대규모 업데이트 예정...

+++ src/main/java/com/ezen/mannamatna/vo/KakaoUserInfoVO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class KakaoUserInfoVO {
 	private int uiNum;
-	private int kuiId;
+	private long kuiId;
 }
